Move WinMDWriter into Writers folder
Move all WinMDWriter partial class files from Generation/ to a new
Writers/ folder with a corresponding Writers namespace, matching the
interop generator's convention of separating writer/emitter types
from the orchestration logic in Generation/.
